+++ stocal/algorithms.py
@@ -23,13 +23,226 @@
 This module provides implementations of different stochastic simulation
 algorithms. All implementations have to implement the TrajectorySampler
 interface by subclassing this abstract base class.
+
+The module also provides several helper classes for general stochastic
+simulation algorithms. DependencyGraph is a generic species-dependency
+graph that gives quick access to the set of affected transitions from
+a set of affected species. MultiDict and PriorityQueue are data
+structures that can be used for managing transitions. MultiDict is
+a mapping from transitions to propensities used in DirectMethod and
+related algorithms. Priority Queue is an indexed priority queue as it
+appears in Gibson-Bruck like NextReactionMethod's. Both data structures
+allow for transitions to be added multiple times, and keep track of
+their multiplicity.
 """
 
 import abc
+import warnings
+from pqdict import pqdict
 
 from ._utils import with_metaclass
 from .structures import multiset
 from .transitions import Event, Reaction
+
+
+class DependencyGraph(dict):
+    """Species-transition dependency graph
+
+    A mapping from species to transitions that are affected by a
+    change in the respecive species' count. A transition counts as
+    affected if it appears among the reactants of the transition.
+    """
+    def add_reaction(self, reaction):
+        """Add reaction to dependencies"""
+        for reactant in reaction.reactants:
+            self.setdefault(reactant, set()).add(reaction)
+
+    def remove_reaction(self, reaction):
+        """Remove reaction from dependencies"""
+        for reactant in reaction.affected_species:
+            if reactant in self:
+                self[reactant].discard(reaction)
+                if not self[reactant]:
+                    del self[reactant]
+
+    def affected_transitions(self, species):
+        """Get all transitions affected by a change in any of the given species
+
+        Species is an iterable and the method returns a set of Transition's.
+        """
+        return set(
+            trans for reactant in species
+            for trans in self.get(reactant, [])
+        )
+
+
+class MultiDict(object):
+    """Dictionary with multiplicity count
+
+    MultiDict supports DirectMethod like samplers with a dictionary
+    that maps transitions to propensities. Unlike normal dictionaries
+    MultiDict keep count of the number of times a key is added, i.e.
+    its multipliciry. Therefore, the signature of MultiDict is
+
+        transition -> (propensity, multiplicity)
+
+    This class and its interface are not module level implementation
+    details and are not specified through tests. They might change in
+    future versions of stocal.
+    """
+    def __init__(self):
+        self._dict = dict()
+
+    def __contains__(self, item):
+        return item not in self._dict
+
+    def add_item(self, key, value):
+        """Add item with associated value.
+
+        If the key has been inserted ealrier, its multiplicity is
+        increased by one. Otherwise, it is stored with multiplicity
+        one."""
+        if key not in self._dict:
+            # insert transition with propensity and multiplicity one
+            self._dict[key] = [value, 1]
+        else:
+            # increase multiplicity count
+            self._dict[key][1] += 1
+
+    def __delitem__(self, key):
+        """Delete all instances of the given key."""
+        del self._dict[key]
+
+    def keys(self):
+        """Return a list of all present transition instances."""
+        warnings.warn("Method will return an iterator in future versions.", DeprecationWarning)
+        return [
+            key for key, (prop, mult) in self._dict.items()
+            for i in range(mult)
+        ]
+
+    def items(self):
+        """Iterate over key, value, multiplicity triples."""
+        for key, item in self._dict.items():
+            value, multiplicity = item
+            yield key, value, multiplicity
+
+    def update_item(self, key, value):
+        """Change value associated with key"""
+        self._dict[key][0] = value
+
+
+class PriorityQueue(object):
+    """Indexed priority queue
+
+    Data structure used for Gibson-Bruck-like transition selection.
+    See the documentation of pqdict for the general properties of the
+    data structure. Unlike the standard indexed priority queue, this
+    implementation allows keys (transitions) to have multiple associated
+    values. In addition, each instance of a key can have optional
+    associated data. The exact mapping is therefore
+
+        key -> [(value_1, data_1), (value_2, data_2), ...]
+
+
+    Instead of assigning values directly, The PriorityQueue constructor
+    takes a function to calculate values for a key. Its signature is
+
+        value_callback(key, data) -> float.
+
+    Custom data is passed as keyword arguments to the add_item method.
+    add_item creates a mutable instance of type PriorityQueue.Item for
+    each transition and binds keyword arguments to it. This data object
+    is retrieved by queue access methods and is also passed to the
+    value_callback.
+    """
+    class Item(object):
+        """Namespace to hold custom associated data"""
+        def __init__(self, **params):
+            for key, value in params.items():
+                setattr(self, key, value)
+
+        def __repr__(self):
+            return '<Data %s>' % ', '.join('%s=%r' % (k, v)
+                                           for k, v
+                                           in self.__dict__.items())
+
+        def __eq__(self, other):
+            return self.__dict__ == other.__dict__
+
+        def __lt__(self, other):
+            return False
+
+
+    def __init__(self, value_callback):
+        self._queue = pqdict()
+        self.value_callback = value_callback
+        self.depletion_value = float('inf')
+
+    def __bool__(self):
+        return bool(self._queue)
+
+    __nonzero__ = __bool__
+
+    def __getitem__(self, key):
+        return self._queue[key]
+
+    def keys(self):
+        """Return a list of all present transition instances."""
+        warnings.warn("Method will return an iterator in future versions.", DeprecationWarning)
+        return [
+            key for key, instances in self._queue.items()
+            for i in instances
+        ]
+
+    def topitem(self):
+        """Yield (value, key, (data,)) triple with lowest value."""
+        key, instances = self._queue.topitem()
+        value, data = instances[0]
+        return value, key, (data,)
+
+    def add_item(self, key, **params):
+        """Add instance of key to the queue.
+
+        Any additional keyword arguments are used to populate a
+        PriorityQueue.Item object that is subsequently passed to
+        the queue's value_callback and stored together with the value
+        of the instance.
+        """
+        if key not in self._queue:
+            self._queue[key] = []
+
+        data = self.Item(**params)
+        value = self.value_callback(key, data)
+        self._queue[key].append((value, data))
+        self._queue[key].sort()
+        self._queue.heapify()
+
+    def remove_item(self, key):
+        """Remove all instances of key with a value of float('inf')"""
+        remaining = [t for t in self._queue[key]
+                     if t != self.depletion_value]
+        if remaining:
+            self._queue[key] = remaining
+        else:
+            del self._queue[key]
+
+    def update_one_instance(self, key):
+        """Recalculates the value for the 'first' instance of key."""
+        instances = self._queue[key]
+        _, data = instances[0]
+        instances[0] = self.value_callback(key, data), data
+        instances.sort()
+        self._queue.heapify()
+
+    def update_items(self, keys):
+        """Recalculate values for all provided keys."""
+        for key in keys:
+            instances = sorted(
+                (self.value_callback(key, data), data)
+                for value, data in self._queue[key]
+            )
+            self._queue[key] = instances
 
 
 class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
@@ -76,10 +289,11 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         self.tmax = tmax
         self.rng = Random(seed)
 
-        for transition in self.process.transitions:
-            self.add_transition(transition)
         self.state = multiset()
         self.update_state(state)
+
+        for transition in self.process.transitions:
+            self.add_transition(transition)
 
     def update_state(self, dct):
         """Modify sampler state.
@@ -122,9 +336,19 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         """
         return float('inf'), None, tuple()
 
+    @abc.abstractproperty
+    def transitions(self):
+        """Return list of all transitions
+
+        Implementations need to return every instance of an added
+        transition, i.e. two copies of a transition that has
+        multiplicity two.
+        """
+        return []
+
     def is_applicable(self, time, transition, *args):
         """True if the transition is applicable
-        
+
         The standard implementation always returns True, i.e. it assumes
         that any transition returned by propose_potential_transition is
         applicable. Overwrite this method if you want to implement
@@ -132,7 +356,7 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         """
         return True
 
-    def perform_transition(self, time, transition):
+    def perform_transition(self, time, transition, *args):
         """Perform the given transition.
 
         Sets sampler.time to time, increases the number of steps,
@@ -152,11 +376,10 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
                 trans.rule = rule
                 self.add_transition(trans)
         self.state += transition.true_products
-        self.prune_transitions()
 
     def reject_transition(self, time, transition, *args):
         """Do not execute the given transition.
-        
+
         The default implementation does nothing. Overwrite this method
         if you, for example, want to prevent the same transition from
         being proposed again.
@@ -190,6 +413,7 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
                 self.reject_transition(time, transition, *args)
             else:
                 self.perform_transition(time, transition, *args)
+                self.prune_transitions()
                 yield transition
 
         if self.step != self.steps and self.tmax < float('inf'):
@@ -218,27 +442,37 @@ class DirectMethod(TrajectorySampler):
             raise ValueError("DirectMethod only works with Reactions.")
         if any(not issubclass(r.Transition, Reaction) for r in process.rules):
             raise ValueError("DirectMethod only works with Reactions.")
-        self.transitions = []
-        self.propensities = []
+        self.dependency_graph = DependencyGraph()
+        self.propensities = MultiDict()
+        self.depleted = []
         super(DirectMethod, self).__init__(process, state, t, tmax, steps, seed)
 
     def add_transition(self, transition):
-        self.transitions.append(transition)
+        self.dependency_graph.add_reaction(transition)
+        propensity = self.calculate_propensity(transition)
+        self.propensities.add_item(transition, propensity)
+
+    def update_state(self, dct):
+        super(DirectMethod, self).update_state(dct)
+        affected_transitions = self.dependency_graph.affected_transitions(dct)
+        self.update_propensities(affected_transitions)
 
     def prune_transitions(self):
-        depleted = [
-            i for i, (p, t) in enumerate(zip(self.propensities, self.transitions))
-            if p == 0. and t.rule
-        ]
-        for i in reversed(depleted):
-            del self.transitions[i]
-            del self.propensities[i]
+        for trans in self.depleted:
+            self.dependency_graph.remove_reaction(trans)
+            del self.propensities[trans]
+        self.depleted = []
+
+    @property
+    def transitions(self):
+        return self.propensities.keys()
 
     def propose_potential_transition(self):
         from math import log
 
-        self.propensities = [r.propensity(self.state) for r in self.transitions]
-        total_propensity = sum(self.propensities)
+        total_propensity = sum(mult*prop
+                               for transition, prop, mult
+                               in self.propensities.items())
         if not total_propensity:
             return float('inf'), None, tuple()
 
@@ -246,12 +480,33 @@ class DirectMethod(TrajectorySampler):
 
         transition = None
         pick = self.rng.random()*total_propensity
-        for propensity, transition in zip(self.propensities, self.transitions):
-            pick -= propensity
+        for transition, prop, mult in self.propensities.items():
+            pick -= mult*prop
             if pick < 0.:
                 break
 
         return self.time + delta_t, transition, tuple()
+
+    def perform_transition(self, time, transition):
+        super(DirectMethod, self).perform_transition(time, transition)
+        affected = self.dependency_graph.affected_transitions(transition.affected_species)
+        self.update_propensities(affected)
+
+    def update_propensities(self, affected_transitions):
+        """Update propensities of all given transitions.
+
+        If the new propensity of a transition is 0 and the transition
+        has been derived by a rule or is an Event, the transition gets
+        added to self.depleted for later pruning."""
+        for trans in affected_transitions:
+            propensity = trans.propensity(self.state)
+            if propensity == 0 and (trans.rule or isinstance(trans, Event)):
+                self.depleted.append(trans)
+            self.propensities.update_item(trans, propensity)
+
+    def calculate_propensity(self, transition):
+        """Return propensity of the given transition for current state."""
+        return transition.propensity(self.state)
 
 
 class FirstReactionMethod(TrajectorySampler):
@@ -297,12 +552,12 @@ class FirstReactionMethod(TrajectorySampler):
         """Returns False for Event's that lack their reactants."""
         if isinstance(transition, Event):
             return transition.reactants <= self.state
-        else :
+        else:
             return super(FirstReactionMethod, self).is_applicable(time, transition, *args)
 
     def reject_transition(self, time, transition, *args):
         """Reject inapplicable Event
-        
+
         Advance system time to transition time and pretend transition
         had happened there, but do not change the state.
         """
@@ -310,7 +565,78 @@ class FirstReactionMethod(TrajectorySampler):
         transition.last_occurrence = time
 
 
-class AndersonNRM(FirstReactionMethod):
+class NextReactionMethod(FirstReactionMethod):
+    """Implementation of Gibson & Bruck's next reaction method.
+
+    A stochastic simulation algorithm, published in
+    M. A. Gibson & J. Bruck, J. Phys. Chem. A 2000, 104, 1876-1889 (1999).
+
+    NextReactionMethod works with processes that feature deterministic
+    transitions, i.e. Event's. It is an optimized version of Gillespie's
+    FirstReactionMethod whose runtime scales logarithmically with the
+    number of reactions.
+
+    See help(TrajectorySampler) for usage information.
+    """
+    def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
+        self.dependency_graph = DependencyGraph()
+        self.firings = PriorityQueue(self.calculate_next_occurrence)
+        self.depleted = []
+        super(FirstReactionMethod, self).__init__(process, state, t, tmax, steps, seed)
+
+    def add_transition(self, transition, **params):
+        self.dependency_graph.add_reaction(transition)
+        self.firings.add_item(transition, **params)
+
+    def update_state(self, dct):
+        super(NextReactionMethod, self).update_state(dct)
+        affected = self.dependency_graph.affected_transitions(dct)
+        self.firings.update_items(affected)
+
+    def prune_transitions(self):
+        for trans in self.depleted:
+            self.dependency_graph.remove_reaction(trans)
+            self.firings.remove_item(trans)
+        self.depleted = []
+
+    @property
+    def transitions(self):
+        return self.firings.keys()
+
+    def propose_potential_transition(self):
+        if self.firings:
+            return self.firings.topitem()
+        else:
+            return float('inf'), None, ()
+
+    def perform_transition(self, time, transition, *args):
+        super(NextReactionMethod, self).perform_transition(time, transition, *args)
+        self.update_firing_times(time, transition, *args)
+
+    def update_firing_times(self, time, transition, *args):
+        """Update next occurrences of firing and all affected transitions"""
+        # update affected firing times
+        affected = self.dependency_graph.affected_transitions(transition.affected_species)
+        self.firings.update_one_instance(transition)
+        self.firings.update_items(affected)
+        # mark depleted reactions
+        for trans in affected:
+            if (any(occurrence[0] == float('inf')
+                    for occurrence in self.firings[trans])
+                    and (trans.rule or isinstance(trans, Event))):
+                self.depleted.append(trans)
+
+    def reject_transition(self, time, transition, *args):
+        self.time = time
+        transition.last_occurrence = time
+        self.firings.update_one_instance(transition)
+
+    def calculate_next_occurrence(self, transition, data):
+        """Calculate next occurrence of given reaction."""
+        return transition.next_occurrence(self.time, self.state, self.rng)
+
+
+class AndersonNRM(NextReactionMethod):
     """Next reaction method modified for time-dependent processes
 
     A stochastic simulation algorithm, published in
@@ -322,47 +648,24 @@ class AndersonNRM(FirstReactionMethod):
 
     See help(TrajectorySampler) for usage information.
     """
-    def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
-        self.T = []
-        self.P = []
-        super(AndersonNRM, self).__init__(process, state, t, tmax, steps, seed)
-
     def add_transition(self, transition):
+        """Add transition with own internal clock (T, P)"""
         from math import log
+        super(AndersonNRM, self).add_transition(
+            transition,
+            T=0, P=-log(self.rng.random())
+        )
 
-        super(AndersonNRM, self).add_transition(transition)
-        self.T.append(0)
-        self.P.append(-log(self.rng.random()))
-
-    def prune_transitions(self):
-        depleted = [
-            k[0] for t, r, k in self.firings
-            if t == float('inf') and (r.rule or isinstance(r, Event))
-        ]
-        for k in reversed(depleted):
-            del self.transitions[k]
-            del self.T[k]
-            del self.P[k]
-
-    def propose_potential_transition(self):
-        def eq_13(trans, target):
-            """Determine timestep for a transition in global time scale"""
-            if isinstance(trans, Event):
-                return trans.next_occurrence(self.time)
-            else:
-                return trans.propensity_meets_target(self.state, self.time, target) +  self.time
-
-        self.firings = [
-            (eq_13(trans, Pk-Tk), trans, (k,))
-            for k, (trans, Pk, Tk)
-            in enumerate(zip(self.transitions, self.P, self.T))
-        ]
-        if self.firings:
-            return min(self.firings, key=lambda item: item[0])
+    def calculate_next_occurrence(self, transition, data):
+        """Determine next firing time of a transition in global time scale"""
+        target = data.P-data.T
+        if isinstance(transition, Event):
+            return transition.next_occurrence(self.time)
         else:
-            return float('inf'), None, tuple()
+            return self.time + transition.propensity_meets_target(
+                self.state, self.time, target)
 
-    def perform_transition(self, time, transition, mu):
+    def perform_transition(self, time, transition, data):
         from math import log
 
         def int_a_dt(trans, delta_t):
@@ -372,7 +675,12 @@ class AndersonNRM(FirstReactionMethod):
             else:
                 return trans.propensity_integral(self.state, self.time, delta_t)
 
-        self.T = [Tk+int_a_dt(trans, time-self.time) for Tk, trans in
-                  zip(self.T, self.transitions)]
-        self.P[mu] -= log(self.rng.random())
+        data.P -= log(self.rng.random())
+        affected = self.dependency_graph.affected_transitions(
+            transition.affected_species
+        )
+        for trans in affected:
+            for time, data in self.firings[trans]:
+                data.T += int_a_dt(trans, time-self.time)
+
         super(AndersonNRM, self).perform_transition(time, transition)