+++ stocal/transitions.py
@@ -81,6 +81,7 @@ class Transition(with_metaclass(abc.ABCMeta, object)):
 
         self.true_reactants = reactants - products
         self.true_products = products - reactants
+        self.affected_species = self.true_reactants.union(self.true_products).domain
 
         self.last_occurrence = -float('inf')
         self._hash = 0
@@ -575,7 +576,7 @@ class Process(object):
         self.transitions = transitions or []
         self.rules = rules or []
 
-    def trajectory(self, state, t=0., tstart=0., tmax=float('inf'), steps=None):
+    def trajectory(self, state, t=0., tstart=0., tmax=float('inf'), steps=None, seed=None):
         """Create trajectory sampler for given state
 
         The method automatically chooses a suitable sampler for the
@@ -593,15 +594,12 @@ class Process(object):
             for rule in self.rules:
                 yield rule.Transition
 
-        # DirectMethod for process with normal reactions
-        if all(issubclass(r, Reaction) and r.is_autonomous
-               for r in transition_types()):
-            from .algorithms import DirectMethod as Sampler
-        # FirstReactionMethod if all reactions are autonomous
-        elif all(r.is_autonomous for r in transition_types()):
-            from .algorithms import FirstReactionMethod as Sampler
-        # AndersonNRM if reactions are non-autonomous
-        else:
+        # select suitable simulation algorithm
+        if any(not r.is_autonomous for r in transition_types()):
+            # AndersonNRM for processes with non-autonomous reactions
             from .algorithms import AndersonNRM as Sampler
+        else:
+            # NextReactionMethod for anything else
+            from .algorithms import NextReactionMethod as Sampler
 
         return Sampler(self, state, tstart, tmax, steps)